requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.1
loguru>=0.7.2
numpy>=1.26.0
scipy>=1.12.0
pydantic>=2.5.0
httpx>=0.27.0
